# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Space DOGE API",
    "version": "1.0.0",
    "description": "Wallet-first Coin Rush Arena with verifiable Scrypt cloud-mining rewards. Every endpoint below reflects the actual current implementation \u2014 request/response shapes were transcribed directly from the route handlers, not hand-guessed.\n\n**Auth**: almost every endpoint requires a session, established via `/api/auth/verify` (SIWE sign-in), which sets an httpOnly `SpaceDOGE_session` cookie. Admin endpoints (`/api/admin/*`) additionally require the session's wallet address to be on the admin allowlist (env `ADMIN_ADDRESSES` or the `AdminUser` table) \u2014 a non-admin session hitting those returns `403`, not `401`.\n\n**Money fields**: all Decimal/Prisma fields are converted to plain JSON `number` before being returned \u2014 never strings.\n\n**Timestamps**: all dates are ISO-8601 strings in responses.",
    "contact": {
      "name": "Space DOGE"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SpaceDOGE_session",
        "description": "httpOnly JWT session cookie set by POST /api/auth/verify."
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static shared secret (env DEPOSIT_CRON_SECRET), used only by the deposit-watch cron trigger."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
    "responses": {
      "NotAuthenticated": {
        "description": "No valid session cookie",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not authenticated"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Validation failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/wallet/deposit-verify": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Verify a user-submitted transaction hash on-chain and credit it immediately if valid",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "txHash",
                  "chainKey"
                ],
                "properties": {
                  "txHash": {
                    "type": "string"
                  },
                  "chainKey": {
                    "type": "string",
                    "example": "BEP20"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Always 200 \u2014 outcome is in the `status` field",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "credited",
                            "pending"
                          ]
                        },
                        "amount": {
                          "type": "number"
                        },
                        "confirmations": {
                          "type": "integer"
                        },
                        "minConfirmations": {
                          "type": "integer"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "not_configured",
                            "invalid_hash",
                            "rpc_error",
                            "not_found",
                            "failed_tx",
                            "no_matching_transfer",
                            "sent_from_different_wallet",
                            "tron_needs_manual_review"
                          ]
                        },
                        "error": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          }
        }
      }
    },
...
    "/api/wallet/convert-pts": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Convert PTS to Game Reward USDT at the fixed platform rate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ptsAmount"
                ],
                "properties": {
                  "ptsAmount": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ptsAmount": {
                      "type": "number"
                    },
                    "usdtAmount": {
                      "type": "number"
                    },
                    "rate": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "402": {
            "description": "Not enough PTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/wallet/withdraw": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Request a withdrawal (creates a PENDING record, debits the ledger immediately)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "destinationAddress",
                  "chain"
                ],
                "properties": {
                  "source": {
                    "type": "string",
                    "enum": [
                      "RECYCLED_USDT",
                      "REFERRAL_USDT"
                    ],
                    "description": "Required for USDT-coin chains; omitted/ignored for the DOGE chain"
                  },
                  "amount": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  },
                  "destinationAddress": {
                    "type": "string"
                  },
                  "chain": {
                    "type": "string",
                    "example": "BEP20"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "RECYCLED_USDT",
                        "REFERRAL_USDT",
                        "AVAILABLE_DOGE"
                      ]
                    },
                    "amount": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING"
                      ]
                    },
                    "destination": {
                      "type": "string"
                    },
                    "chain": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, unavailable chain, missing source, below minimum, or bad address format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "402": {
            "description": "Not enough balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mining/simulate-reward-demo": {
      "post": {
        "tags": [
          "Mining"
        ],
        "summary": "Dev-only: credit simulated Game Reward USDT (blocked in production)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1000,
                    "exclusiveMinimum": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "simulated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "403": {
            "description": "Not available in production",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}